Overhaul build system; Update core and plugins
Most significant changes listed below

* Remove files generated by autotools
* Add bootstrap script
* Allow users to build motoko without warden support
* Add compatibility layer
* Add INTERNAL_REQUEST interface
* Add Include directive for plugin configuration
* Add CD key rotation capability to plugin mcp game
* Improve plugin chicken reliability
* Add socket detection to plugin pickit (broken)
* Fix several minor bugs
* Update configuration file templates

+++ lib/util/file.c
@@ -54,7 +54,7 @@ size_t file_read(const char *file, byte *buf, size_t len) {
 		free(e);
 		return -1;
 	}
-	size_t tmp, total = 0;
+	int tmp, total = 0;
 	do {
 		tmp = read(fd, buf + total, len - total);
 		if (tmp < 0) {
@@ -85,7 +85,7 @@ size_t file_write(const char *file, byte *buf, size_t len) {
 		free(e);
 		return -1;
 	}
-	size_t tmp, total = 0;
+	int tmp, total = 0;
 	do {
 		tmp = write(fd, buf + total, len - total);
 		if (tmp < 0) {
